requirements.txt
-i https://pypi.org/simple
babel==2.7.0
bcrypt==3.1.7
blinker==1.4
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.0
flask-babelex==0.9.3
flask-bcrypt==0.7.1
flask-images==3.0.1
flask-login==0.4.1
flask-mail==0.9.1
flask-oauth==0.12
flask-principal==0.4.0
flask-security==3.0.0
flask-sqlalchemy==2.4.0
flask-wtf==0.14.2
flask==1.0.3
google-api-core==1.13.0
google-auth-oauthlib==0.4.0
google-auth==1.6.3
google-cloud-bigquery==1.15.0
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
markupsafe==1.1.1
numpy==1.16.4
oauth2==1.9.0.post1
oauthlib==3.0.1
pandas-gbq==0.10.0
pandas==0.24.2
passlib==1.7.1
pillow==6.0.0
pillowcase==2.0.0
protobuf==3.8.0
pyasn1-modules==0.2.5
pyasn1==0.4.5
pycparser==2.19
pydata-google-auth==0.1.3
python-dateutil==2.8.0
pytz==2019.1
requests-oauthlib==1.2.0
requests==2.22.0
rsa==4.0
six==1.12.0
speaklater==1.3
sqlalchemy==1.3.5
urllib3==1.25.3
werkzeug==0.15.4
wtforms==2.2.1
gunicorn==19.9.0
